+++ Changelog.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.2.8
+
+ * Fixed a regression with meeting without end from 0.2.7
+ * Updated the json importer to format version 3
+ * The json importer now has a better heuristic for determining the date of a paper
+
 ## 0.2.7
 
  * `manage.py import_json` now by defaults send notifications, this can be disabled with `--no-notify-users`